Asuite: add PREUPLOAD.cfg and release_checker
The release_checker will check all requirements are fulfiled before
uploading prebuilts. If insufficient permission are detected, auto
correctifying the perms if an affirmative answer is given.

This CL is only checking the permission status, and later on we'll
support running integration tests or other necessary tests via the
repohook.

bug: 128805741

Test: git add binaries without executable permission, and run
      repo upload
      An warning message will be given for the detected issue.

Change-Id: I62afda7c8800f4851e7c566b421c771e12dd4986