ASuite: check uploaded files and run smoke test if available.
This CL checks if the PREUPLOAD_FILES fulfils the requirements, and run
smoke test accordingly if the test script is available.

release_checker.py checks 4 states:
*If the uploaded file is a known prebuilt(atest, aidegen, acloud), it
must be executable.
*If the uploaded file is not a knwon prebuilt but locate along with them,
e.g. atest/linux-x86/atest-test, is must be not executable to prevent
from contaminating PATH.
*If the uploaded files are known prebuilts but more than 1 in quantity,
prompt the user to upload one at a time and exit.
*If the uploaded files is a smoke test script, it must be executable.

The filename of the smoke test script is always named 'smoke_tests'
without suffix because we wanna give users the flexiblity of writing
smoke test in any language. If the uploaded file is atest/linux-x86/atest
and atest/smoke_tests is available and executable, the release_checker.py
will invoke the script during repo upload.

Bug: 138827336

Test: source build/envsetup.sh; lunch; cd prebuilts/asuite
      chmod a-x atest/linux-x86/atest aidegen/linux-x86/aidegen
      git add atest/linux-x86/atest aidegen/linux-x86/aidegen
      ./release_checker.py
      # will see "insufficient permission" message.

      chmod a+x atest/linux-x86/atest aidegen/linux-x86/aidegen
      git add atest/linux-x86/atest aidegen/linux-x86/aidegen
      ./release_checker.py
      # will see "upload one binary at a time" message.

      cp atest/linux-x86/atest atest/linux-x86/atest-dev
      git add atest/linux-x86/atest-dev
      ./release_checker.py
      # will see "Only atest, acloud and aidegen are allowed."

      echo -e '#!/usr/bin/env bash\nfake(){ echo "Test fail."; return 1; };fake' >atest/smoke_tests
      chmod +x atest/smoke_tests
      git add atest/linux-x86/atest
      ./release_checker.py
      # will see "Test fail".
      ./release_checker.py -n
      # won't see any error.
      rm atest/smoke_tests

Change-Id: Idcf8c36810ffaa419d5c7580d661f955f5c37c7b